+++ NEWS.md
@@ -4,6 +4,7 @@
 - Removed support for deprecated nesting in `@nn_network` slot
 
 ## Added
+- New `gefToGiotto()` interoperability function to convert gef object from Stereo-seq to giotto
 - New `giottoToAnnData()` interoperability function to convert giotto object to squidpy flavor anndata .h5ad file(s)
 - Add missing `create_nn_net_obj()` internal constructor function for S4 `nnNetObj`
 - New `createGiottoMerscopeObject()` convenience function